Prevent truncation of ID offset to 32 bits

+++ src/internal.hpp
@@ -1484,7 +1484,7 @@ struct Internal {
       if (clause_id > (uint64_t) opts.lratorigclscount) clause_id -= opts.lratsolvercount;
       // In order to avoid clashing with any prior solvers computing with the same solver ID,
       // begin your own clauses at an according offset (but still with the right remainder!)
-      clause_id += opts.lratskippedepochs * opts.lratsolvercount * 1e8;
+      clause_id += ((uint64_t) opts.lratskippedepochs) * opts.lratsolvercount * 1e8;
     }
     // Go to next clause ID
     clause_id += opts.lratsolvercount;
